requirements.txt
flask
vanna[postgres,chromadb,zhipuai]
db-dtypes
python-dotenv